Fix parsing of --distribution plane=<plane_size> option.
Parsing of the "plane" distribution option didn't properly check if the
syntax was completely correct. For example, if there was an equals sign
immediately after the -m then Slurm assumed that it was a plane
distribution and then parsed whatever string came after that as the
size. For example, "-m=8gb" was interpreted as plane distribution of
size 8. This commit ensures that the correct syntax must be used for
the "plane" distribution type.

Since no other distribution type is valid with the "plane" distribution,
this commit checks for "plane" as the first characters in the
distribution string. If anything comes after the plane size, it is
ignored. For example, in the string "plane=5:block", ":block" is ignored.

Bug 9248